Deleted Dashboard and moved MyCourses to Profile folder
